@@ -1,17 +1,22 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 import './index.css'
-import App from './p1-main/m1-ui/u1-app/App'
-import * as serviceWorker from './serviceWorker'
+import App from './s1-main/App'
+import reportWebVitals from './reportWebVitals'
+import store from './s2-homeworks/hw10/bll/store'
+import {Provider} from 'react-redux'
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
+root.render(
     <React.StrictMode>
-        <App/>
-    </React.StrictMode>,
-    document.getElementById("root")
+        {/*для дз 10*/}
+        <Provider store={store}>
+            <App/>
+        </Provider>
+    </React.StrictMode>
 )
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister()
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals()
